requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
idna==3.10
packaging==24.2
PyYAML==6.0.2
requests==2.32.3
urllib3==2.2.3
zstandard==0.23.0
